requirements.txt
streamlit>=1.36
pandas>=2.1
plotly>=5.18
gspread>=5.12
oauth2client>=4.1
python-dateutil>=2.9
pyarrow>=16.0
